+++ alicloud/extension_cms.go
@@ -1,0 +1,16 @@
+package alicloud
+
+const (
+	Average = "Average"
+	Minimum = "Minimum"
+	Maximum = "Maximum"
+)
+
+const (
+	MoreThan        = ">"
+	MoreThanOrEqual = ">="
+	LessThan        = "<"
+	LessThanOrEqual = "<="
+	Equal           = "=="
+	NotEqual        = "!="
+)